Moved Ship out of GameObject.ts and into its own file
